chore: アプリからの MUI 直 import を ESLint で止め、対応表を単一ソース化
DS 準拠率 100% は今のところ運用で守られているだけで、次のプロダクトで
崩れる。仕組みで守る。

- `apps/*/src` と `src/pages` に no-restricted-imports を追加。
  DS に同等品がある MUI 部品（TextField / Chip / Card / Dialog 等 24 名）
  の直 import を error で止める
- DS 本体 (src/components) と Storybook の story は対象外。DS は MUI を
  包む側なので当然 MUI を import するし、story は比較のために直に使う
- Box / Grid / Stack / Typography 等のレイアウト原始要素は DS に同等品が
  無いので対象外

対応表を scripts/ds-equivalents.mjs に切り出して単一ソースにした。
ESLint と計測 (pnpm ds:adoption) が同じ表を見るので、DS に部品を足した
ときに片方だけ更新されて「計測では未準拠なのに lint は通る」状態に
ならない。

エラーメッセージは名前ごとに 1 エントリにした。まとめて 1 つの message に
すると対応表 20 件分が毎回出て、肝心の置き換え先が読み取れなかった
（実際に一度そうなったので直した）。

検証:
- 現状のリポジトリ全体で 0 件（アプリ・LP・DS 本体すべて）
- アプリに `import { TextField, Chip } from '@mui/material'` を仕込むと
  2 件とも error になり、それぞれ CustomTextField / CustomChip を
  名指しすることを確認
- 同じファイルで Box / Typography だけなら通ることを確認
- DS 本体 (customChip.tsx) が MUI を直 import しても通ることを確認
- pnpm lint / test:run 522 passed / ds:adoption 100%
